requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
langchain==0.0.346
langchain-community==0.0.10
openai==0.28.0
google-api-python-client==2.108.0
gspread==5.11.0
oauth2client==4.1.3
python-dotenv==1.0.0
pydantic==2.5.0
python-multipart==0.0.6